Graphical GWO algorithm fo finding nearest point
This code was written to better understand the Gray Wolf algorithm. The gray wolf algorithm is one of the optimization algorithms in the field of metaheuristic algorithms. In this code, the target point and the wolves are drawn graphically and as desired, and during the repetitions, the wolves try to reach the target point. For more information, you can visit our Persian website. http://programstore.ir/